src/postgres_db_interface.py
import os
import psycopg2
from sqlalchemy import create_engine

class PostgresDBInterface:
    """
    This class provides an interface to connect to a PostgreSQL database and execute basic operations.
    """

    def __init__(self, username, password, hostname, database):
        """
        The constructor for the PostgresDBInterface class.

        :param username: str, username for the PostgreSQL server
        :param password: str, password for the PostgreSQL server
        :param hostname: str, hostname for the PostgreSQL server
        :param database: str, database name in the PostgreSQL server
        """
        self.username = username
        self.password = password
        self.hostname = hostname
        self.database = database
        self.engine = None

    def connect(self):
        """
        Connects to the PostgreSQL database using provided credentials.

        :returns: None
        :raises Exception: If the connection was not successful
        """
        try:
            connection_string = f"postgresql+psycopg2://{self.username}:{self.password}@{self.hostname}/{self.database}"
            self.engine = create_engine(connection_string)
            # Check the connection
            connection = self.engine.connect()
            print("Successfully connected to the database.")
            connection.close()
        except psycopg2.OperationalError as e:
            if 'password authentication failed' in str(e):
                raise Exception("Invalid username or password.")
            elif 'database' in str(e) and 'does not exist' in str(e):
                raise Exception("The specified database does not exist.")
            elif 'could not translate host name' in str(e):
                raise Exception("The specified host could not be resolved.")
            else:
                raise Exception("An error occurred while trying to connect to the database.")
        except Exception as e:
            raise Exception("An unexpected error occurred while trying to connect to the database.")


    def disconnect(self):
        """
        Closes the connection to the PostgreSQL database.
        """
        pass

    def execute_query(self, query):
        """
        Executes a given SQL query on the PostgreSQL database.
        """
        pass

    def fetch_data(self, query):
        """
        Executes a given SQL query and returns the resulting data from the PostgreSQL database.
        """
        pass

    def inject_anomaly(self, table_name, anomaly_definition):
        """
        Injects an anomaly into a given table in the PostgreSQL database according to the provided anomaly definition.
        """
        pass

    def profile_database(self):
        """
        Profiles the PostgreSQL database, gathering and returning information about its schemas, tables, and columns.
        """
        pass

    def verify_anomaly(self, table_name, anomaly_definition):
        """
        Verifies if an anomaly has been correctly injected into a given table in the PostgreSQL database.
        """
        pass

    def _translate_query(self, query):
        """
        Translates a given SQL query to the PostgreSQL dialect.
        """
        pass
